Refactor tests
- Enhance SpyOnPointerClickHandler with ClickTimestamps, WasClicked, and ClickCount properties
- Use ClickTimestamps.Count for ClickCount property (no redundant data storage)
- Remove duplicate SpyPointerClickHandler and SpySequenceTrackingClickHandler classes
- Update all test code to use unified SpyOnPointerClickHandler
- Maintain backward compatibility with existing log functionality

Benefits:
- Single data source eliminates inconsistency risks
- Memory efficient design with computed properties
- Full feature coverage: logging, counting, timing, and flags
- Simplified maintenance with reduced code duplication
