+++ Tests/Runtime/TestDoubles/SpyOnPointerClickHandler.cs
@@ -1,6 +1,8 @@
-﻿// Copyright (c) 2023 Koji Hasegawa.
+﻿// Copyright (c) 2023-2025 Koji Hasegawa.
 // This software is released under the MIT License.
 
+using System;
+using System.Collections.Generic;
 using System.Runtime.CompilerServices;
 using UnityEngine;
 using UnityEngine.EventSystems;
@@ -8,16 +10,44 @@ using UnityEngine.EventSystems;
 namespace TestHelper.Monkey.TestDoubles
 {
     /// <summary>
-    /// Pointer click event handler
+    /// Spy implementation of IPointerClickHandler that records click events with timestamps, counts, and logging.
     /// </summary>
     [AddComponentMenu("/")] // Hide from "Add Component" picker
     internal class SpyOnPointerClickHandler : MonoBehaviour, IPointerClickHandler
     {
+        /// <summary>
+        /// Gets the list of timestamps when OnPointerClick was called.
+        /// </summary>
+        public List<DateTime> ClickTimestamps { get; private set; } = new List<DateTime>();
+
+        /// <summary>
+        /// Gets a value indicating whether OnPointerClick was called.
+        /// </summary>
+        public bool WasClicked => ClickTimestamps.Count > 0;
+
+        /// <summary>
+        /// Gets the number of times OnPointerClick was called.
+        /// </summary>
+        public int ClickCount => ClickTimestamps.Count;
+
+        /// <summary>
+        /// Resets the spy state.
+        /// </summary>
+        public void Reset()
+        {
+            ClickTimestamps.Clear();
+        }
+
         private void Log([CallerMemberName] string member = "")
         {
             Debug.Log($"{this.name}.{member}");
         }
 
-        public void OnPointerClick(PointerEventData eventData) => Log();
+        /// <inheritdoc />
+        public void OnPointerClick(PointerEventData eventData)
+        {
+            Log();
+            ClickTimestamps.Add(DateTime.Now);
+        }
     }
 }
