--- Tests/Runtime/TestDoubles/SpyPointerClickHandler.cs.meta
@@ -1,2 +1,0 @@
-fileFormatVersion: 2
-guid: 950af100a0c62490e8ca635168b69374
