--- Tests/Runtime/TestDoubles/SpySequenceTrackingClickHandler.cs.meta
@@ -1,2 +1,0 @@
-fileFormatVersion: 2
-guid: 182fa339cd7df4a84a2330bf20979d54
