+++ Tests/Scenes/Canvas.unity.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 836d367b7934240fea373eb3180d51eb
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
